requirements.txt
Django==2.2.12
djangorestframework==3.10.3
requests==2.23.0
